// MyCompany/Angular/src/app/models/inExpendDetail.ts
export class inExpendDetail {
    IDDetail: number;
    IDExpend: number;
    IDProduct: number;
    ProductName: string;
    Volume: number;
    //Price: number;
    BuyPrice: number;
    SellPrice: number;
    BuyAmount: number;
    SellAmount: number;
    IDFaktur: string;
    IDSupplier: number;
    CurrencySell: string;
    CurrencyBuy: string;
    CurrencySellAmount: string;
    CurrencyBuyAmount: string;
    Created: Date;
    StrCreated: string;
}